requirements.txt
aiohttp
aiosqlite
anthropic
beautifulsoup4
discord.py==2.5.2
feedparser
python-dotenv

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
freezegun>=1.2.0